Added progress bars, reorganized files
