+++ tibiawikisql/__init__.py
@@ -1,1 +1,1 @@
-__all__ = ["utils", "__main__"]
+from . import parsers
