+++ tibiawikisql/parsers/__init__.py
@@ -1,0 +1,13 @@
+from . import achievements
+from . import common
+from . import creatures
+from . import database
+from . import houses
+from . import imbuements
+from . import items
+from . import map
+from . import utils
+from . import npcs
+from . import quests
+from . import spells
+
