+++ tibiawikisql/parsers/database.py
