+++ tibiawikisql/parsers/map.py
