+++ tibiawikisql/parsers/utils.py
